+++ demo/app/index.js
@@ -12,6 +12,8 @@ const demos = {
   linkedTimelines: require('./demo-linked-timelines').default,
   elementResize: require('./demo-element-resize').default,
   plugins: require('./demo-plugins').default,
+  stickyHeader: require('./demo-sticky-header').default,
+  customItems: require('./demo-cusom-items').default,
   stickyHeader: require('./demo-sticky-header').default
 }