+++ src/lib/row/GroupRows.js
@@ -1,11 +1,15 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
+import GroupRow from './GroupRow'
 
-export default class HorizontalLines extends Component {
+export default class GroupRows extends Component {
   static propTypes = {
     canvasWidth: PropTypes.number.isRequired,
     lineCount: PropTypes.number.isRequired,
-    groupHeights: PropTypes.array.isRequired
+    groupHeights: PropTypes.array.isRequired,
+    onRowClick: PropTypes.func.isRequired,
+    onRowDoubleClick: PropTypes.func.isRequired,
+    clickTolerance: PropTypes.number.isRequired
   }
 
   shouldComponentUpdate(nextProps) {
@@ -17,24 +21,30 @@ export default class HorizontalLines extends Component {
   }
 
   render() {
-    const { canvasWidth, lineCount, groupHeights } = this.props
+    const {
+      canvasWidth,
+      lineCount,
+      groupHeights,
+      onRowClick,
+      onRowDoubleClick,
+      clickTolerance
+    } = this.props
     let lines = []
 
-    var totalHeight = 0
     for (let i = 0; i < lineCount; i++) {
       lines.push(
-        <div
+        <GroupRow
+          clickTolerance={clickTolerance}
+          onClick={evt => onRowClick(evt, i)}
+          onDoubleClick={evt => onRowDoubleClick(evt, i)}
           key={`horizontal-line-${i}`}
-          className={i % 2 === 0 ? 'rct-hl-even' : 'rct-hl-odd'}
+          isEvenRow={i % 2 === 0}
           style={{
-            top: `${totalHeight}px`,
-            left: '0px',
             width: `${canvasWidth}px`,
             height: `${groupHeights[i] - 1}px`
           }}
         />
       )
-      totalHeight += groupHeights[i]
     }
 
     return <div className="rct-horizontal-lines">{lines}</div>